fix: task test

+++ src/routes/tasks/tasks.handlers.ts
@@ -4,24 +4,35 @@ import * as HttpStatusPhrases from "stoker/http-status-phrases";
 
 import type { AppRouteHandler } from "@/lib/types";
 
-import db from "@/db";
 import { tasks } from "@/db/schema";
 import { ZOD_ERROR_CODES, ZOD_ERROR_MESSAGES } from "@/lib/constants";
 
 import type { CreateRoute, GetOneRoute, ListRoute, PatchRoute, RemoveRoute } from "./tasks.routes";
 
 export const list: AppRouteHandler<ListRoute> = async (c) => {
+  const db = c.get("db");
+  if (!db) {
+    throw new Error("db is not available");
+  }
   const tasks = await db.query.tasks.findMany();
   return c.json(tasks);
 };
 
 export const create: AppRouteHandler<CreateRoute> = async (c) => {
+  const db = c.get("db");
+  if (!db) {
+    throw new Error("db is not available");
+  }
   const task = c.req.valid("json");
   const [inserted] = await db.insert(tasks).values(task).returning();
   return c.json(inserted, HttpStatusCodes.OK);
 };
 
 export const getOne: AppRouteHandler<GetOneRoute> = async (c) => {
+  const db = c.get("db");
+  if (!db) {
+    throw new Error("db is not available");
+  }
   const { id } = c.req.valid("param");
   const task = await db.query.tasks.findFirst({
     where(fields, operators) {
@@ -42,6 +53,10 @@ export const getOne: AppRouteHandler<GetOneRoute> = async (c) => {
 };
 
 export const patch: AppRouteHandler<PatchRoute> = async (c) => {
+  const db = c.get("db");
+  if (!db) {
+    throw new Error("db is not available");
+  }
   const { id } = c.req.valid("param");
   const updates = c.req.valid("json");
 
@@ -82,6 +97,10 @@ export const patch: AppRouteHandler<PatchRoute> = async (c) => {
 };
 
 export const remove: AppRouteHandler<RemoveRoute> = async (c) => {
+  const db = c.get("db");
+  if (!db) {
+    throw new Error("db is not available");
+  }
   const { id } = c.req.valid("param");
   const result = await db.delete(tasks)
     .where(eq(tasks.id, id));
